refactor: 프로젝트 구조 재정립
