+++ src/features/lost/components/ItemCard.styles.ts
