+++ src/features/lost/components/Registration.styles.ts
