+++ src/features/lost/modal/ModalNotification.styles.ts
