+++ src/features/lost/modal/ModalPost.styles.ts
